Initial Commit
A very quick and dirty script programmed in 8 hours. You will need to adjust the api_key value, and may need to change the value on line 79; the input volume on line 30.

This script can only currently speak on MacOS, if you would like windows support, change the code on line 111.